Move and rename Decoder class
Renamed class from Ima to IMAADPCMDecoder to have a more specific name
Moved the class to a Helper folder as Extension namespace is used only for ObjectType
